lambda_deploy/requirements_lambda.txt
# Lambda-specific requirements
# Note: Some packages may need to be pre-compiled for Lambda

# Core OCR libraries
pytesseract==0.3.13
Pillow==10.0.1  # Use older version for Lambda compatibility
opencv-python-headless==4.8.1.78  # Use headless version for Lambda
easyocr==1.7.2

# Image processing
numpy==1.24.3  # Use older version for Lambda compatibility
scikit-image==0.21.0
imutils==0.5.4

# PDF processing
PyPDF2==3.0.1
pdf2image==1.16.3

# Machine Learning (for EasyOCR)
torch==2.0.1  # Use older version for Lambda compatibility
torchvision==0.15.2

# Utilities
python-dotenv==1.0.0